+++ src/analyze/src/components/MainNav.vue
@@ -17,10 +17,14 @@
          class="navbar-menu"
          :class="{'is-active': isMobileMenuOpen}">
       <div class="navbar-start">
-        <router-link to="/model"
-            class="navbar-item has-dropdown is-hoverable">
-          <a class="navbar-link" href="/model">
-            Model
+
+        <router-link to="/" class="navbar-item navbar-child">
+          Files
+        </router-link>
+
+        <div class="navbar-item has-dropdown is-hoverable">
+          <a class="navbar-link">
+            Analyze
           </a>
           <div class="navbar-dropdown
                 is-boxed"
@@ -38,34 +42,21 @@
             </router-link>
             </template>
           </div>
+        </div>
+
+        <router-link to="/dashboards"
+        class="navbar-item navbar-child">
+          Dashboards
         </router-link>
-        <router-link to="/extract"
-          class="navbar-item">
-          Extract
-        </router-link>
-        <router-link to="/load"
-          class="navbar-item">
-          Load
-        </router-link>
-        <router-link to="/transform"
-          class="navbar-item">
-          Transform
-        </router-link>
-        <a class="navbar-item" href="/">
-          Analyze
-        </a>
-        <a class="navbar-item disabled" disabled="true" href="/">
-          Notebook
-        </a>
-        <a class="navbar-item" href="/orchestrations">
-          Orchestrate
-        </a>
+
       </div>
       <div class="navbar-end">
+
         <router-link to="/settings"
         class="navbar-item navbar-child">
           Settings
         </router-link>
+
       </div>
     </div>
   </nav>