+++ src/analyze/src/filters/pretty.js
@@ -1,0 +1,7 @@
+export default function pretty(value) {
+  try {
+    return JSON.stringify(JSON.parse(value), null, 2);
+  } catch (e) {
+    return value;
+  }
+}