+++ src/analyze/src/store/index.js
@@ -2,6 +2,7 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 import repos from './modules/repos';
 import designs from './modules/designs';
+import dashboards from './modules/dashboards';
 import settings from './modules/settings';
 import orchestrations from './modules/orchestrations';
 
@@ -13,6 +14,7 @@ export default new Vuex.Store({
   modules: {
     repos,
     designs,
+    dashboards,
     settings,
     orchestrations,
   },